Fix nits in OpenWhisk minio documentation
The `jq` command listed resulted in two separate JSON dictionaries in
the output as opposed to modifying the `address` field (change `,` to
`|`) and had missing quotes around the address.

In addition, this commit rewords the reason for changing the address
field to be a bit more explicit and gives a more concrete example (in a
different IP prefix than the internal address for clarity) of how to
determine the right address.